Module rename and documentation
Renamed one module extra-routes to help-view and added "quickstart" for evaluators to try modules in a Karaf from CLI
