+++ help-view/pom.xml
@@ -9,8 +9,8 @@
         <version>1.0.0-SNAPSHOT</version>
     </parent>
 
-    <artifactId>extra-routes</artifactId>
-    <name>Bundle for Vaadin application with extra routes</name>
+    <artifactId>help-view</artifactId>
+    <name>Vaadin view for help</name>
     <packaging>jar</packaging>
 
     <dependencies>
@@ -18,7 +18,7 @@
         <dependency>
             <groupId>com.example</groupId>
             <artifactId>project-base-karaf</artifactId>
-            <version>1.0-SNAPSHOT</version>
+            <version>1.0.0-SNAPSHOT</version>
         </dependency>
         
     </dependencies>
