+++ help-view/src/main/java/com/example/helpview/HelpView.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.example.routes;
+package com.example.helpview;
 
 import com.example.starter.base.ContentLayout;
